move referee handling to rules and make configurable
