+++ FSFV.Gameplanner.UI/FSFV.Gameplanner.UI/appsettings.json
@@ -43,5 +43,11 @@
     },
     "FooterDateFormat": "dd.MM.yyyy",
     "GameStartTimeFormat": "HH:mm"
+  },
+
+  "RefereeUpdate": {
+      "E": {
+          "HasReferees": false
+      }
   }
 }
